chore(logger): code refactoring
