+++ src/util/index.ts
@@ -1,3 +1,3 @@
-export * from './logger';
+export { logger } from './logger';
 export * from './download';
 export * from './extract';
